Patterns stage logic moved to WorkflowBase.
